+++ corda-testacles-mocknetwork/src/main/kotlin/com/github/manosbatsis/corda/testacles/mocknetwork/MockNetworkHelper.kt
@@ -90,13 +90,18 @@ open class MockNetworkHelper(
         return NodeHandles(mockNetwork, mockNodes)
     }
 
+    /** Apply any custom config if one was given for it's package */
+    protected open fun TestCordapp.maybeWithConfig(packageName: String) =
+            mockNetworkConfig.cordappPackageConfigs[packageName]
+                    ?.let { this.withConfig(it) }
+                    ?: this
 
     /** Initialize, but not start, the network */
     protected open fun buildMockNetwork(): MockNetwork {
         val cordapps = mutableListOf<TestCordapp>()
         // Add project's cordapp classes, if configured
-        mockNetworkConfig.cordappProjectPackage?.also {
-            cordapps.add(cordappWithPackages(it))
+        mockNetworkConfig.cordappProjectPackage?.also { pkg ->
+            cordapps.add(cordappWithPackages(pkg).maybeWithConfig(pkg))
         }
         // Add any JAR cordapps based on packages configured
         mockNetworkConfig.cordappPackages
@@ -105,7 +110,7 @@ open class MockNetworkHelper(
                     it == mockNetworkConfig.cordappProjectPackage
                             || it.isBlank()
                 }
-                .mapTo(cordapps) { TestCordapp.findCordapp(it) }
+                .mapTo(cordapps) { TestCordapp.findCordapp(it).maybeWithConfig(it) }
         return MockNetwork(MockNetworkParameters(
                 cordappsForAllNodes = cordapps,
                 threadPerNode = mockNetworkConfig.threadPerNode,