+++ docs/changelog.md
@@ -5,6 +5,11 @@ The following sections describe major changes per version
 and can be helpful with version upgrades.
 
 
+## 0.19
+
+- Added Cordapp Config support to MockNetworkConfig
+- Added NodeParamsHelper to use alongside NodeDriverHelper
+
 ## 0.18
 
 - Support Corda OS/CE 4.7